Translate radio buttons into RadioGroup component
This commit does a bunch of stuff I'm really proud of. The RadioGroup component is now abstract and builds collections of radio buttons based on the question it's passed. I moved the questions from QuizContainer and into Quiz during the build, but will probably move them back in a future commit. RadioGroup should also probably be broken out into it's own file.