begin building the quizReducer
quizReducer will handle all the state information for the quiz. Currently, it's broken. =( We need to figure out why cuisineNames.concat is wiping the object property. Other than that, we connected the quizContainer component to the store, as I want it to handle all the communication with Redux. It's the 'smart' component, eventually, everything it renders will be 'dumb' components.